requirements.txt
facenet-pytorch==2.5.3
torch==2.3.0
torchvision==0.18.0
onnx==1.16.0
onnxruntime==1.18.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
fastapi==0.111.0
uvicorn==0.30.1
aiofiles==23.2.1